#780 Wicket `@SpringBean` injection generates illegal access warning on recent JDKs
- some interfaces
