+++ inception-kb/src/test/java/de/tudarmstadt/ukp/inception/kb/KnowledgeBaseExporterTest.java
@@ -57,7 +57,7 @@ import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.clarin.webanno.support.JSONUtil;
-import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBaseProperties;
+import de.tudarmstadt.ukp.inception.kb.config.KnowledgeBasePropertiesImpl;
 import de.tudarmstadt.ukp.inception.kb.exporter.KnowledgeBaseExporter;
 import de.tudarmstadt.ukp.inception.kb.model.KnowledgeBase;
 
@@ -97,7 +97,7 @@ public class KnowledgeBaseExporterTest
 
         when(schemaService.listAnnotationFeature(sourceProject)).thenReturn(features(sourceProject));
 
-        sut = new KnowledgeBaseExporter(kbService, new KnowledgeBaseProperties(), schemaService);
+        sut = new KnowledgeBaseExporter(kbService, new KnowledgeBasePropertiesImpl(), schemaService);
     }
 
     @Test
